+++ esbuild.config.mjs
@@ -1,6 +1,7 @@
 import esbuild from 'esbuild';
 import process from 'process';
 import builtins from 'builtin-modules';
+import svg from 'esbuild-plugin-svg';
 
 const isProduction = process.argv.includes('production');
 
@@ -13,6 +14,7 @@ const commonOptions = {
     ],
     format: 'cjs', // Obsidian plugins use CommonJS
     treeShaking: true,
+    plugins: [svg()],
 };
 
 const buildOptions = {
@@ -30,5 +32,5 @@ if (isProduction) {
     esbuild.context(buildOptions).then(context => {
         console.log('Watching for changes...');
         context.watch();
-    }).catch(() => process.exit(1));
+    }).catch(() => process.exit.bind(1));
 }