+++ jest.config.mjs
@@ -8,9 +8,9 @@ export default {
   testRegex: 'tests/.*\\.test\\.ts$',
   moduleNameMapper: {
     '^obsidian$': '<rootDir>/tests/__mocks__/obsidian.ts',
+    '\\.svg$': '<rootDir>/tests/__mocks__/svg-mock.ts',
   },
   collectCoverageFrom: [
-    'src/logic/**/*.ts',
-    'src/components/**/*.ts',
+    'src/**/*.ts',
   ],
 };